io.IncrementalNewlineDecoder is not a subclass of codecs.IncrementalDecoder
The documentation says that this class "inherits codecs.IncrementalDecoder",
but it's not true for the C implementation.

The python implementation, _pyio.IncrementalNewlineDecoder, *does*
inherit from codecs.IncrementalDecoder, but that's not the implementation
that you get with `from io import IncrementalNewlineDecoder`.

cpython has a ticket for this issue here: https://github.com/python/cpython/issues/75903

In the meantime, I think this is more correct.

Related to https://github.com/python/typeshed/issues/3968