bs4: Make `Tag.attrs` more permissive (#11487)
Make `Tag.attrs` more permissive to be in line with actual code

`Tag.attrs` is a dict can technically have `list[str]` as a valid value, not just `str`. Making the value type a union with `Any` allows `list[str]` to be valid in a type checker's eyes if a user narrows it.

Based on discussion in, and fixes, #8755.