Update overload in stdlib/sqlite3/dbapi2.pyi file.
As per @srittau's comment in #11571: The first overload handles the default `factory` but we can't have a positional arg without a default following an positional arg with a default. "To work around this, we actually need to split this overload into two, one for providing the preceding arguments positionally, and one for providing the factory as keyword argument".

Co-authored-by: Sebastian Rittau <srittau@rittau.biz>