Tighten annotation of logging.getLevelName
To better reflect the implementation's behaviour,
https://github.com/python/typeshed/pull/2730 changed
`logging.getLevelName` to accept `int | str` and return `Any` (the
latter due to the need to avoid union return types).  However, this
isn't ideal if you're passing in an `int`, in which case the
implementation always returns a `str`.  Add overloads for this.

This is all arguably a bit unfortunate in light of
https://github.com/python/typeshed/issues/1842#issuecomment-360929192,
but I don't want to relitigate that here.  I've at least left a comment.