datetime: use more precise typing for fromtimestamp()
The `fromtimestamp()` API differs in behaviour between `date` and `datetime`.

In the case of `date`, the parameter is passed directly to `time.localtime()`
or `time.gmtime()`, which only accept integral values, so it is obvious for
it to take `SupportsInt` (despite the `DeprecationWarning` when passing, for
instance, `Decimal` numbers.) Due to the nature of the type, sub-second
precision is not necessary.

In the case of `datetime`, sub-second precision *is* necessary, so at least
up to Python 3.12, it's this implementation that extracts the decimal part
of the parameter. But it then take the remainder integral part and passes it
to the same `time.localtime()`, which requires a conversion to integer.

This means that while the function takes `float` parameters just fine, it
requires them to also support `int` conversion — so it will not accept e.g.
`Fraction` parameters.

    >>> datetime.date.fromtimestamp(fractions.Fraction(1, 1))
    Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    TypeError: an integer is required (got type Fraction)
    >>> datetime.datetime.fromtimestamp(decimal.Decimal(3))
    datetime.datetime(1969, 12, 31, 16, 0, 3)

It also doesn't quite always preserve the sub-second precision for non-float
types:

    >>> t = time.time()
    >>> t
    1718884427.5039804
    >>> d = decimal.Decimal(t)
    >>> d
    Decimal('1718884427.5039803981781005859375')
    >>> datetime.datetime.fromtimestamp(t)
    datetime.datetime(2024, 6, 20, 4, 53, 47, 503980)
    >>> datetime.datetime.fromtimestamp(d)
    datetime.datetime(2024, 6, 20, 4, 53, 47)