Enforce handling `sys.std{in,out,err} is None`
As per the Python docs [1], these streams can be None (see [2] for a
list of possible circumstances). The current usage of MaybeNone causes a
false negative, which is a shame because spreading the little known
knowledge that `sys.stderr.write()` is broken code is something that
type checkers would otherwise be ideal at.

[1]: https://docs.python.org/3/library/sys.html#sys.__stderr__
[2]: https://github.com/python/cpython/issues/122633?notification_referrer_id=NT_kwDOAdgearQxMTc4OTE2ODgyNzozMDk0MDc3OA#issuecomment-2267146355