move various cases of 'stub does not have *args argument' to "don't fix"
I fixed _threading_local.local.__new__ instead of just moving it.
That one was also hiding a 'stub argument "cls" should be positional-only' message,
and when I looked at the implementation, the use of *args, **kwargs seemed
intended to allow for genuine functionality if someone wanted to create
their own subclass. It seemed reasonable to just add it.