adjusting socketserver.BaseServer
This one might be a bad idea. fileno, get_request, and server_bind
are documented to exist on BaseServer, but they don't. Of these,
only get_request is actually called from the existing implementation
of BaseServer (as part of handle_request and serve_forever).

In the ideal case, they'd probably all be abstract methods in
CPython.