fix the weird special builtins
These were moved to "don't fix" before stubs for _sitebuiltins
were added to typeshed. We have the actual classes now; does it
cause any problems to use them?

Includes a couple small fixes to the classes in _sitebuiltins,
which sync them up with the signature of the function version
that this MR replaces.