pyserial: make serial.threaded.ReaderThread generic in the type of Protocol
This change makes `serial.threaded.ReaderThread` generic in the type of the `serial.threaded.Protocol` that it wraps.

Currently, `ReaderThread`'s `connect` and context-manager `__enter__` methods, along with the `protocol` instance property, return a `serial.threaded.Protocol`. However, it is usually the case that `ReaderThread` will wrap a specific subclass of `serial.threaded.Protocol`, which has additional protocol-specific methods defined on it. So the return value of `serial.threaded.Protocol` is not really useful and it needs to be cast to the actual `serial.threaded.Protocol` subclass type, so that other protocol methods can be called without causing a type-checker error.

This change makes `ReaderThread` generic, and the wrapped `serial.threaded.Protocol` type a TypeVar, so that `ReaderThread` methods can return the specific `serial.threaded.Protocol` subclass rather than `serial.threaded.Protocol` itself.