stdlib: add overloads for itertools.product() up to an arity of 15
Before this change, the maximum usable arity for itertools.product() was 10
(increased in 2e83e654b18a, #12023, which raised it from 6 to 10).

Fixes #13490, albeit partially: just based on this change, a user has no way of
knowing if he is hitting this limit or not.

This discoverability problem will probably be best addressed with changes in
documentation or in the type checkers.