Future's type var now default to Any
This makes it possible to write `Future` instead of `Future[Any]` to
annotate something as accepting any `Future`. This reduces visual
clutter and avoids the need to comment on each `Future[Any]` per our
documentation policy. It also allows us to distinguish between the
"any Future allowed" case (bare `Future`) and the "a Future with a
value we can't express" case (`Future[Any]`). This mirrors the
annotations for `type`.