Fix `shared` type in `_interpreters`
It is a `dict[str, Any]` or `None`:

```python
>>> import _interpreters
>>> i = _interpreters.create()

>>> _interpreters.exec(i, 'print(1)', shared=False)
Traceback (most recent call last):
  File "<python-input-3>", line 1, in <module>
    _interpreters.exec(i, 'print(1)', shared=False)
    ~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: _interpreters.exec() argument 3 must be dict, not bool
```

Non-`str` keys:

```python
>>> _interpreters.exec(i, 'print(1)', shared={1: 1})
Traceback (most recent call last):
  File "<python-input-4>", line 1, in <module>
    _interpreters.exec(i, 'print(1)', shared={1: 1})
    ~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: bad argument type for built-in operation
```

Real sharing:

```python
>>> _interpreters.exec(i, 'print(1)', shared={'a': 1})
1
```

Some objects can't be shared. Currently only a few can be. But, I think that restricting this type yet is not a good idea, because more can be added in the future. Also, `(1, 2)` is ok, `(1, object())` is not