Make Mapping.get(default) less constrained
At the moment we say that `default` can either be the same as the value type, or anything it wants to be. Given:

```
def f(mapping: Mapping[str, str]):
    reveal_type(mapping.get("x", "y"))
```

The `_T` is entirely unconstrained, and thus `str | int` is a perfectly valid type. This change makes it obviously `str | str` which collapses to `str` correctly.

This fix makes Pyrefly happier, although Pyright already seems to deal with it, but this doesn't harm Pyright.