Allow docutils.readers.standalone.Reader.document to be a docutils.nodes.document
The use of this in docutils is not trivial but in testing I changed `document` to a property and created
a setter and found that https://github.com/docutils/docutils/blob/9f04802532e77542c8edf6b7f4bf78bab141cf6e/docutils/docutils/readers/__init__.py#L103
was setting this as a `nodes.document`.
Intuitively it likely is not always None, as well :)

I am making this change because Sphinx sets this to a `nodes.document`.