Remove obsolete comments from turtle.pyi
While it would be possible to use the hack described in the comment
now that pytype support has been removed, that hack is a bit too
"magical" for stubs that are supposed to be fairly straight-forward
API descriptions. This would also introduce extra stub-only attributes
and would imply that e.g. `_Screen.setup()` and `setup()` refer to the
same object at runtime when they don't.