Update jsonschema stubs to track latest changes
In python-jsonschema/jsonschema#1396 , the type signature for
`Validator.__init__` is updated to better match the runtime signature.
This backports the fix to typeshed, keeping the copies of this data in
sync. python-jsonschema/jsonschema#1396 is, itself, a response to
feedback on `jsonschema` about the changes in `typeshed` #14327.

In addition to the `__init__` fix, a couple of additional small changes
are made, both in `jsonschema` and here in the stubs:

1. In `jsonschema`, the type for `create()` in `validators.py` was
   updated to be notated with `-> type[Validator]`. This was necessary
   for internal testing on types to correctly read that validator
   classes created by this factory implement the protocol.

2. Here, in order to better guarantee that the types align, the
   `_Validator` class (which does not exist in `jsonschema`, but is only
   defined here in the stubs) now inherits from `Validator`.

3. The init signature for `_Validator` is updated to match

4. `tests/mypy_test.py` flags the `schema` instance variable annotation
   as mismatching between `Validator` and `_Validator`. Review against
   the `jsonschema` source reveals that `_Validator` was closer to
   correct, so `Validator` is fixed to match.

Any further changes (e.g., elimination of `_Validator` or changing
`create`'s return type annotation) are left as potential future work.