Rework xml.etree to avoid typevars
This is an alternative to #14674

While investigating, I also noticed that QName classes can be the value
of both Element.tag and Eelement.text, as well in the result of
Element.items() (and thus as keys and values in Element.attrib).
The test suite in cpython also includes explict tests for assigning
None to Element.tag as a form of suppressing it, so I added that into
the annotation as well.