correct grammatical error in CONTRIBUTING.md to "in their own right"
The pronoun and its noun did not agree in grammatical number. Furthermore, the previous author seems to have started writing the idiomatic phrase "in its own right", and then gotten confused and completed it by writing the end of the idiomatic phrase "on its own merits". The former better expresses the concept under discussion here (which could also be paraphrased as "themselves", "in-and-of-themselves", or "also", although less precisely), so I've gone with it.